PRO-2950: move Executor under Axn::Core
