+++ lib/axn/async/exception_reporting.rb
@@ -85,7 +85,7 @@ module Axn
             instance = action_class.send(:new, **_deserialize_job_args(job_args))
             # Apply the same inbound preprocessing/defaults a normal `.call` would, so defaulted /
             # preprocessed inputs resolve as the worker saw them (not the raw job args).
-            Axn::Executor.new(instance).prepare_inbound_for_facets!
+            Axn::Core::Executor.new(instance).prepare_inbound_for_facets!
             {
               tags: Core::Tagging.resolve(action_class._tags, action: instance, from: :inputs),
               dimensions: Core::Tagging.resolve(action_class._dimensions, action: instance, from: :inputs),
