+++ spec/axn/core/messages_spec.rb
@@ -771,7 +771,7 @@ RSpec.describe Axn do
     shared_examples "action with custom error layers" do
       before do
         # Verify exception handling triggers via Executor
-        expect_any_instance_of(Axn::Executor).to receive(:trigger_on_exception).and_call_original
+        expect_any_instance_of(Axn::Core::Executor).to receive(:trigger_on_exception).and_call_original
       end
 
       it { expect(result).not_to be_ok }
